refactor(core): promote pkg/health to internal/core/health (no behavior change)
