+++ internal/core/health/watch.go
